+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.2.0
+
+ * Feature: Support for Mongoose connections
+ * Breaking change: Deprecated "connectionOpts" in favor of "options"
+
 # 3.1.0
 
  * Feature: Added caching feature